UTH-118: render tab contexts
